Test refactoring - moved `add_user_to_site` to `tests.helpers`
This is so that other tests don't have to include `test_metrics` to use
this convenience method
