+++ tests/helpers.py
@@ -6,6 +6,8 @@ from packaging import version
 
 import organizations
 
+import figures.sites
+
 
 def make_course_key_str(org, number, run='test-run', **kwargs):
     '''
@@ -30,8 +32,22 @@ def organizations_support_sites():
     return orgs_has_site
 
 
+def add_user_to_site(user, site):
+    orgs = figures.sites.get_organizations_for_site(site)
+    assert orgs.count() == 1, 'requires a single org for the site. Found {}'.format(
+        orgs.count())
+    if not organizations.models.UserOrganizationMapping.objects.filter(
+        user=user, organization=orgs.first()).exists():
+        obj, created = organizations.models.UserOrganizationMapping.objects.get_or_create(
+            user=user,
+            organization=orgs[0],
+        )
+        return obj, created
+
+
 def django_filters_pre_v1():
     """Returns `True` if the installed Django Filters package is before '1.0.0'
     """
     import django_filters
     return version.parse(django_filters.__version__) < version.parse('1.0.0')
+
